+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Version 1.1.0 (2017-07-??)
 
 * [new] Add basic utilities for field/method reflection.
+* [new] Add the ability to create a default instance of a class.
+* [new] Add utilities to work with types.
+* [new] Add utilities to work with priorities.
 * [chg] Reflection performance improvements.
 
 # Version 1.0.2 (2017-04-29)